+++ print_kiosk/s3_status.py
@@ -4,7 +4,7 @@ from botocore.exceptions import NoCredentialsError
 import yaml
 import time
 from common.common import load_config
-from uploader.upload_to_s3 import wait_for_network_connection
+from common.common import wait_for_network_connection
 
 def get_keys(key_path):
     with open(key_path, "r") as key_file:
@@ -44,17 +44,18 @@ class S3Status:
         url = f"https://{self.bucket_name}.s3.{location}.amazonaws.com/{file_name}"
         return url
 
-wait_for_network_connection()
-uploader = S3Status()
-config = load_config("print_config")
-file_path = config["status_file_path"]
-file_content = ""
+if __name__ == "__main__":
+    wait_for_network_connection()
+    uploader = S3Status()
+    config = load_config("print_config")
+    file_path = config["status_file_path"]
+    file_content = ""
 
-while True:
-    with open(file_path, "r") as file_obj:
-        data = file_obj.read()
-    if data != file_content:
-        file_content = data
-        url = uploader.upload_file(file_path)
-        print(url)
-    time.sleep(10)
+    while True:
+        with open(file_path, "r") as file_obj:
+            data = file_obj.read()
+        if data != file_content:
+            file_content = data
+            url = uploader.upload_file(file_path)
+            print(url)
+        time.sleep(10)
