+++ uploader/upload_to_s3.py
@@ -1,9 +1,6 @@
-import yaml
 import qrcode
-import glob
 import time
 import os
-import socket
 import time
 from datetime import datetime
 from common.common import load_config
@@ -12,33 +9,7 @@ from uploader.photo_service import PhotoService
 from uploader.google_photos_upload import GooglePhotos
 from uploader.smugmug import SmugMug
 from uploader.s3_photos import S3Photos
-
-
-def check_network_connection(host="8.8.8.8", port=53, timeout=3):
-    """
-    Check network connectivity by trying to connect to a specific host and port.
-    Google's public DNS server at 8.8.8.8 over port 53 (DNS) is used as default.
-    """
-    try:
-        socket.setdefaulttimeout(timeout)
-        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        sock.connect((host, port))
-        sock.close()
-        return True
-    except socket.error as ex:
-        print(f"Network is not reachable. Error: {ex}")
-        return False
-
-
-def wait_for_network_connection():
-    """
-    Wait indefinitely until the network is available.
-    """
-    print("upload_to_s3.py: Waiting for network connection...")
-    while not check_network_connection():
-        time.sleep(5)  # wait for 5 seconds before checking again
-    print("upload_to_s3.py: Network connection established.")
-
+from common.common import wait_for_network_connection
 
 def create_qr_code(url, qr_code_file_path):
     """
